Completed my Schema for Cassandra and Seeded it with 100,000 records
