+++ cassandra/import_airbnb_listing.err.20190606_041755
@@ -1,0 +1,1 @@
+id;city;title;hostImage;roomInfo;numberOfGuests;numberOfBedrooms;numberOfBeds;numberOfBaths;isSuperhost;isGreatLocation;isSparklingClean;isGreatCheckIn;isSelfCheckIn;description;amenities_basic;amenities_dining;amenities_bedAndBath;sleepingArrangements
