+++ cassandra/import_airbnb_listing.err.20190606_042116
@@ -1,0 +1,1 @@
+111;new york;a great place;http://google.com/myimag;cozy and warm;8;4;8;3;false;false;true;true;true;oh my god, it is such a unique place to be in!;[\"wifi\"\",\"\"laptop area\"\",\"\"iron\"\"]\";['kithcen','utensil'];['shampoo','hair dryer','hanger'];{'Bedroom 1':'2 queen beds','Bedroom 2':'1 king bed','Common Space':'2 sofa beds'}
