+++ package.json
@@ -28,6 +28,8 @@
     "babel-loader": "^8.0.6",
     "babel-preset-env": "^1.7.0",
     "css-loader": "^2.1.1",
+    "csv-write-stream": "^2.0.0",
+    "csv-writer": "^1.3.0",
     "enzyme": "^3.9.0",
     "enzyme-adapter-react-16": "^1.13.1",
     "eslint": "^5.16.0",
@@ -47,6 +49,7 @@
     "cors": "^2.8.5",
     "express": "^4.16.4",
     "faker": "^4.1.0",
+    "fs": "0.0.1-security",
     "install": "^0.12.2",
     "jquery": "^3.4.1",
     "mongoose": "^5.5.9",
